Create utility for map exports
